requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-CORS==3.0.10

openai==1.6.1
python-dotenv==1.0.0
Werkzeug==2.2.2
Jinja2==3.0.3
SQLAlchemy==2.0.41
aiohttp==3.9.1
pymysql==1.1.0 